finished the vec.c functions still needs work in them
