+++ include/mat.h
@@ -1,0 +1,7 @@
+
+typedef struct MATRIX4 GRFDP;
+
+struct MATRIX4
+{
+
+};
