+++ src/CMakeLists.txt
@@ -21,10 +21,10 @@ set (FILES
 	# ...
  "keyhole.cpp")
 
-add_library(DECOMP STATIC ${FILES} "keyhole.cpp") # todo: use shared
+add_library(DECOMP STATIC ${FILES} "keyhole.cpp" "vec.cpp" "so.cpp" "dialog.cpp" "speaker.cpp" "mat.cpp") # todo: use shared
 
 target_include_directories(DECOMP PUBLIC ${SLY1_INCLUDE_DIR})
 
-add_executable(main main.cpp "keyhole.cpp")
+add_executable(main main.cpp "keyhole.cpp" "vec.cpp" "so.cpp" "dialog.cpp" "speaker.cpp" "mat.cpp")
 
 target_link_libraries(main PUBLIC DECOMP)
