+++ src/dialog.cpp
@@ -1,0 +1,1 @@
+#include <dialog.h>
