+++ src/speaker.cpp
@@ -1,0 +1,1 @@
+#include <speaker.h>
